feat: Adjusts LessonController response and LessonService data treatment to improve API usability

+++ src/main/java/com/thalesbensi/CoursesManagementAPI/services/LessonService.java
@@ -2,7 +2,9 @@ package com.thalesbensi.CoursesManagementAPI.services;
 
 import com.thalesbensi.CoursesManagementAPI.dto.Lesson.LessonDTO;
 import com.thalesbensi.CoursesManagementAPI.exceptions.ResourceNotFoundException;
+import com.thalesbensi.CoursesManagementAPI.model.Course;
 import com.thalesbensi.CoursesManagementAPI.model.Lesson;
+import com.thalesbensi.CoursesManagementAPI.repositories.CourseRepository;
 import com.thalesbensi.CoursesManagementAPI.repositories.LessonRepository;
 import jakarta.validation.Valid;
 import org.springframework.stereotype.Service;
@@ -13,11 +15,18 @@ import java.util.List;
 public class LessonService {
 
     private final LessonRepository lessonRepository;
-    public LessonService(LessonRepository lessonRepository) {this.lessonRepository = lessonRepository;}
+    private final CourseRepository courseRepository;
+
+    public LessonService(LessonRepository lessonRepository, CourseRepository courseRepository) {
+        this.lessonRepository = lessonRepository;
+        this.courseRepository = courseRepository;
+    }
 
     public List<LessonDTO> getAllLessons() {
-        List<Lesson> lessons = lessonRepository.findAll();
-        return lessons.stream().map(LessonDTO::fromEntity).toList();
+        return lessonRepository.findAll()
+                .stream()
+                .map(LessonDTO::fromEntity)
+                .toList();
     }
 
     public LessonDTO getLessonById(@Valid Long id) {
@@ -26,24 +35,40 @@ public class LessonService {
         return LessonDTO.fromEntity(lesson);
     }
 
-    public LessonDTO createLesson(@Valid Lesson lesson) {
-        lessonRepository.save(lesson);
-        return LessonDTO.fromEntity(lesson);
+    public LessonDTO createLesson(@Valid LessonDTO lessonDTO) {
+        Course course = courseRepository.findById(lessonDTO.course().id())
+                .orElseThrow(() -> new ResourceNotFoundException("Course not found for lesson creation"));
+
+        Lesson lesson = new Lesson();
+        lesson.setTitle(lessonDTO.title());
+        lesson.setDescription(lessonDTO.description());
+        lesson.setUrlVideo(lessonDTO.urlVideo());
+        lesson.setCourse(course);
+
+        Lesson savedLesson = lessonRepository.save(lesson);
+        return LessonDTO.fromEntity(savedLesson);
     }
 
-    public LessonDTO updateLesson(@Valid Lesson lesson, Long id) {
+    public LessonDTO updateLesson(@Valid LessonDTO lessonDTO, Long id) {
         Lesson lessonToBeUpdated = lessonRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Lesson with ID " + id + " not found! :("));
-        lessonToBeUpdated.setTitle(lesson.getTitle());
-        lessonToBeUpdated.setDescription(lesson.getDescription());
-        lessonToBeUpdated.setUrlVideo(lesson.getUrlVideo());
-        lessonToBeUpdated.setCourse(lesson.getCourse());
-        lessonRepository.save(lessonToBeUpdated);
-        return LessonDTO.fromEntity(lessonToBeUpdated);
+
+        Course course = courseRepository.findById(lessonDTO.course().id())
+                .orElseThrow(() -> new ResourceNotFoundException("Course not found for lesson update"));
+
+        lessonToBeUpdated.setTitle(lessonDTO.title());
+        lessonToBeUpdated.setDescription(lessonDTO.description());
+        lessonToBeUpdated.setUrlVideo(lessonDTO.urlVideo());
+        lessonToBeUpdated.setCourse(course);
+
+        Lesson updatedLesson = lessonRepository.save(lessonToBeUpdated);
+        return LessonDTO.fromEntity(updatedLesson);
     }
 
     public void deleteLesson(@Valid Long id) {
+        if (!lessonRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Lesson with ID " + id + " not found! :(");
+        }
         lessonRepository.deleteById(id);
     }
-
 }
